+++ google_play_scraper/__init__.py
@@ -3,4 +3,4 @@ from .features.app import app
 from .features.reviews import reviews, reviews_all
 from .features.permissions import permissions
 
-VERSION = __version__ = "1.0.4"
+VERSION = __version__ = "1.0.5"