+++ google_play_scraper/constants/request.py
@@ -15,8 +15,10 @@ class Format(ABC):
 
 class Formats:
     class _Detail(Format):
-        URL_FORMAT = "{}/store/apps/details?id={{app_id}}&hl={{lang}}&gl={{country}}".format(
-            PLAY_STORE_BASE_URL
+        URL_FORMAT = (
+            "{}/store/apps/details?id={{app_id}}&hl={{lang}}&gl={{country}}".format(
+                PLAY_STORE_BASE_URL
+            )
         )
 
         def build(self, app_id: str, lang: str, country: str) -> str:
@@ -26,8 +28,10 @@ class Formats:
             return None
 
     class _Reviews(Format):
-        URL_FORMAT = "{}/_/PlayStoreUi/data/batchexecute?hl={{lang}}&gl={{country}}".format(
-            PLAY_STORE_BASE_URL
+        URL_FORMAT = (
+            "{}/_/PlayStoreUi/data/batchexecute?hl={{lang}}&gl={{country}}".format(
+                PLAY_STORE_BASE_URL
+            )
         )
 
         def build(self, lang: str, country: str) -> str:
@@ -60,8 +64,10 @@ class Formats:
             return result.encode()
 
     class _Permissions(Format):
-        URL_FORMAT = "{}/_/PlayStoreUi/data/batchexecute?hl={{lang}}&gl={{country}}".format(
-            PLAY_STORE_BASE_URL
+        URL_FORMAT = (
+            "{}/_/PlayStoreUi/data/batchexecute?hl={{lang}}&gl={{country}}".format(
+                PLAY_STORE_BASE_URL
+            )
         )
 
         def build(self, lang: str, country: str) -> str: