+++ google_play_scraper/features/app.py
@@ -29,9 +29,17 @@ def app(app_id: str, lang: str = "en", country: str = "us") -> Dict[str, Any]:
     result = {}
 
     for k, spec in ElementSpecs.Detail.items():
-        content = spec.extract_content(dataset)
+        if isinstance(spec, list):
+            for sub_spec in spec:
+                content = sub_spec.extract_content(dataset)
 
-        result[k] = content
+                if content is not None:
+                    result[k] = content
+                    break
+        else:
+            content = spec.extract_content(dataset)
+
+            result[k] = content
 
     result["appId"] = app_id
     result["url"] = url